requirements.txt
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Session==0.6.0
Bootstrap-Flask==2.3.3
flask-bcrypt==1.0.1
flask_cors==4.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==21.2.0
boto3==1.34.103
pandas==2.2.2
rauth==0.7.3
s3fs==2024.12.0
llama-index==0.12.14
openai==1.60.2
stripe==11.6.0
PyJWT==2.8.0
cryptography==42.0.5
platformdirs==4.3.6
pytz==2024.1